wj
wj

+++ bbq-db-project/src/main/java/com/bbq/db/project/model/Book.java
@@ -10,7 +10,7 @@ public class Book {
 	private String author;
 	private String isbn;
 	private int quantity;
-	private double price;
+	private int price;
 	private String publisher;
 	private Date publishTime;
     private String pic;
@@ -32,7 +32,7 @@ public class Book {
 	public int getQuantity() {
 		return quantity;
 	}
-	public double getPrice() {
+	public int getPrice() {
 		return price;
 	}
 	public String getPublisher() {
@@ -59,7 +59,7 @@ public class Book {
 	public void setQuantity(int quantity) {
 		this.quantity = quantity;
 	}
-	public void setPrice(double price) {
+	public void setPrice(int price) {
 		this.price = price;
 	}
 	public void setPublisher(String publisher) {
@@ -69,7 +69,7 @@ public class Book {
 		this.publishTime = publishTime;
 	}
 	public Book(int bookId, User user, String title, String author,
-			String isbn, int quantity, double price, String publisher,
+			String isbn, int quantity, int price, String publisher,
 			Date publishTime) {
 		super();
 		this.bookId = bookId;
